refactor: unify playable list under useTableColumnVisibility
usePlayableListColumnVisibility was a parallel implementation of the
same idea: valid columns + defaults + storage key + a special always-
visible column. The only divergent piece was responsive hiding
(configurable only at md+ breakpoints), which now lives in
useTableColumnVisibility behind a 'responsive' flag on the config.

- Move the playable-list column data into playableListColumnConfig in
  config/tables.ts (responsive: true, alwaysVisible: ['title']).
- Extend useTableColumnVisibility to accept the 'responsive' flag and
  expose isConfigurable() so the action menu can still gate its
  visibility checkboxes on screen size.
- Switch PlayableListItem, PlayableListHeader, and PlayableListHeader-
  ActionMenu to useTableColumnVisibility(playableListColumnConfig).
- Delete usePlayableListColumnVisibility.ts and its spec.
