+++ resources/assets/js/config/tables.ts
@@ -21,3 +21,39 @@ export const artistTableColumnConfig = {
   defaultColumns: readonly ArtistTableColumnName[]
   alwaysVisible: readonly ArtistTableColumnName[]
 }
+
+export const playableListColumnConfig = {
+  storageKey: 'playable-list-columns',
+  validColumns: [
+    'track',
+    'genre',
+    'year',
+    'title',
+    'artist',
+    'album',
+    'duration',
+    'play_count',
+    'rating',
+    'favorite',
+    'playlist_collaborator',
+    'playlist_added_at',
+  ] as const,
+  defaultColumns: [
+    'track',
+    'title',
+    'artist',
+    'album',
+    'duration',
+    'favorite',
+    'playlist_collaborator',
+    'playlist_added_at',
+  ] as const,
+  alwaysVisible: ['title'] as const,
+  responsive: true,
+} satisfies {
+  storageKey: string
+  validColumns: readonly PlayableListColumnName[]
+  defaultColumns: readonly PlayableListColumnName[]
+  alwaysVisible: readonly PlayableListColumnName[]
+  responsive: boolean
+}
